pyproject.toml
[project]
name = "mystery-recipe-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi",
  "uvicorn",
  "pydantic",
  "pydantic-settings",
  "pymongo",
  "httpx",
  "python-dotenv",
  "PyJWT",
  "openai",
  "clerk-backend-api>=3.0.3",
]
